refactor tests dir structure
